backend/requirements.txt
fastapi
uvicorn
python-binance
sqlalchemy
psycopg2-binary
pandas
numpy
pydantic[email] # Add email validation dependency
python-dotenv
websockets
PyJWT[crypto]
cryptography